requirements.txt
langgraph
langchain
langchain-community
langchain-google-genai
requests
faiss-cpu
tiktoken
networkx
pyvis
streamlit
python-dotenv
